Simplify app and move relationship tests
